backend/requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-CORS>=4.0.0
Flask-JWT-Extended>=4.5.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
Werkzeug>=2.3.0
requests>=2.28.0
bcrypt>=4.0.0
schedule>=1.2.0
google-genai>=0.2.0